[subsystems] New chapter for subsystem related topics
Some more detailed info about adapting Sailfish OS to things
where kernel APIs are used directly or where some configuration
is needed for the system to work properly. A new "subsystems" chapter
was added for this.

Some topics there were added as TODOs as they are needed but I don't
have the details of how the adaptation goes there.

Signed-off-by: Kalle Jokiniemi <kalle.jokiniemi@jolla.com>